Move PCFFromMatrix to its own file
