+++ ramanujan/pcf/__init__.py
@@ -1,3 +1,4 @@
-from .pcf import PCF, PCFFromMatrix
+from .pcf import PCF
+from .pcf_from_matrix import PCFFromMatrix
 
 __all__ = ["PCF", "PCFFromMatrix"]
