+++ ramanujan/pcf/pcf_test.py
@@ -2,7 +2,7 @@ from pytest import approx
 from sympy.abc import c, n
 
 from ramanujan import Matrix
-from ramanujan.pcf import PCF, PCFFromMatrix
+from ramanujan.pcf import PCF
 
 
 def test_repr():
@@ -57,28 +57,3 @@ def test_deflate_all():
     b_n = n**2 - c * n
     pcf = PCF(c_n * a_n, c_n.subs(n, n - 1) * c_n * b_n)
     assert PCF(a_n, b_n) == pcf.deflate_all()
-
-
-def test_pcf_from_matrix():
-    from ramanujan.cmf.known_cmfs import cmf1, c0, c1, c2, c3
-
-    cmf = cmf1.subs([[c0, 0], [c1, 1], [c2, 1], [c3, 3]])
-    matrix = cmf.trajectory_matrix([1, 1], [1, 1])
-    assert PCFFromMatrix(matrix).pcf == PCF(5 + 10 * n, 1 - 9 * n**2)
-
-
-def test_pcf_from_matrix_parametric():
-    from ramanujan.cmf.known_cmfs import cmf1, c0, c1, c2, c3
-
-    cmf = cmf1.subs([[c0, 0], [c1, 1], [c2, 1], [c3, c]])
-    matrix = cmf.trajectory_matrix([1, 1], [1, 1])
-    assert PCFFromMatrix(matrix).pcf == PCF((1 + 2 * n) * (c + 2), 1 - (c * n) ** 2)
-
-
-def test_pcf_from_matrix_relative_limit():
-    from ramanujan.cmf.known_cmfs import cmf1, c0, c1, c2, c3
-
-    cmf = cmf1.subs([[c0, 0], [c1, 1], [c2, 1], [c3, c]])
-    matrix = cmf.trajectory_matrix([1, 1], [1, 1])
-    pcf = PCFFromMatrix(matrix)
-    assert pcf.relative_limit() == Matrix([[2, 1], [0, 1]])
